# api.yaml
openapi: 3.1.0
info:
  title: github-engineer-dashboard
  version: 0.1.0
paths:
  /healthz:
    get:
      tags:
      - meta
      summary: Healthz
      operationId: healthz_healthz_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
  /api/rate-limit:
    get:
      tags:
      - github
      summary: Rate Limit
      operationId: rate_limit_api_rate_limit_get
      parameters:
      - name: x-github-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Github-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimit'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{username}/activity:
    get:
      tags:
      - github
      summary: User Activity
      operationId: user_activity_api_users__username__activity_get
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
          title: Username
      - name: x-github-token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Github-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserActivity'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Health:
      properties:
        status:
          type: string
          title: Status
        version:
          type: string
          title: Version
      type: object
      required:
      - status
      - version
      title: Health
    RateLimit:
      properties:
        limit:
          type: integer
          title: Limit
        remaining:
          type: integer
          title: Remaining
        used:
          type: integer
          title: Used
        reset:
          type: integer
          title: Reset
      type: object
      required:
      - limit
      - remaining
      - used
      - reset
      title: RateLimit
    UserActivity:
      properties:
        username:
          type: string
          title: Username
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        public_repos:
          type: integer
          title: Public Repos
        followers:
          type: integer
          title: Followers
        following:
          type: integer
          title: Following
        event_counts:
          additionalProperties:
            type: integer
          type: object
          title: Event Counts
        total_events:
          type: integer
          title: Total Events
        cached:
          type: boolean
          title: Cached
          default: false
      type: object
      required:
      - username
      - public_repos
      - followers
      - following
      - event_counts
      - total_events
      title: UserActivity
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError